Renaming user-interaction files
