+++ src/user_interaction/mod.rs
@@ -1,2 +1,2 @@
-pub mod render_seconds_human_readable;
-pub mod convert_input_to_seconds;
+pub mod human_readable;
+pub mod convert_input;
